Run drjacoby MCMC

Run flexible MCMC through drjacoby.

run_mcmc(
  data,
  df_params,
  misc = list(),
  loglike,
  logprior,
  burnin = 1000,
  samples = 10000,
  rungs = 1,
  chains = 5,
  coupling_on = TRUE,
  GTI_pow = 3,
  cluster = NULL,
  pb_markdown = FALSE,
  silent = FALSE
)

Arguments

data
a vector of data values. When using C++ versions of the likelihood and/or prior these values are treated internally as doubles, so while integer and boolean values can be used, keep in mind that these will be recast as doubles in the likelihood (i.e. TRUE = 1.0).

df_params
a dataframe of parameters. Must contain the following elements:

name - the parameter name.

min - the minimum value of the parameter. -Inf is allowed.

max - the maximum value of the parameter. Inf is allowed.

init - the initial value of the parameter.

misc
optional list object passed to likelihood and prior.

loglike
TODO.

logprior
TODO.

burnin
the number of burn-in iterations.

samples
the number of sampling iterations.

rungs
the number of temperature rungs used in Metropolis coupling (see coupling_on).

chains
the number of independent replicates of the MCMC chain to run. If a cluster object is defined then these chains are run in parallel, otherwise they are run in serial.

coupling_on
whether to implement Metropolis-coupling over temperature rungs.

GTI_pow
the power used in the generalised thermodynamic integration method. Must either be a single positive integer or a vector of integers with length equal to the number of rungs.

cluster
option to pass in a cluster environment, allowing chains to be run in parallel (see package "parallel").

pb_markdown
whether to run progress bars in markdown mode, meaning they are only updated when they reach 100 being printed to markdown files.

silent
whether to suppress all console output.