Run MCMC

Run MCMC.

run_mcmc(data, df_params, loglike, logprior, burnin = 1000,
  samples = 10000, rungs = 1, chains = 1, burnin_phases = 3,
  bw_update = TRUE, cov_update = c(FALSE, TRUE, TRUE),
  coupling_on = TRUE, GTI_pow = 3, cluster = NULL,
  pb_markdown = FALSE, silent = FALSE)

Arguments

data
TODO.

df_params
a dataframe of parameters. Must contain the following elements: TODO

loglike
TODO.

logprior
TODO.

burnin
the number of burn-in iterations (see also burnin_phases).

samples
the number of sampling iterations.

rungs
the number of temperature rungs used in Metropolis coupling (see coupling_on).

chains
the number of independent replicates of the MCMC chain to run. If a cluster object is defined then these chains are run in parallel, otherwise they are run in serial.

burnin_phases
the number of times a burn-in should be carried out prior to entering the sampling phase. Several parameters can be defined separately for each phase of the burn-in, giving greater control over which tuning methods are applied, and in which order. For example, by default the proposal bandwidth is updated in the first burn-in phase, followed by two burn-in phases in which both the proposal bandwidth and covariance are updated. The complete list of arguments that can be specified separately for each burn-in phase is: burnin, bw_update, cov_update, coupling_on. Each of these arguments can be specified as a vector of length burnin_phases, or as a single value in which case the same value applies over every phase.

bw_update
whether the proposal bandwidth should be updated dynamically via Robbins-Monro (see also burnin_phases).

cov_update
whether the proposal covariance matrix should be updated dynamically from the posterior draws (see also burnin_phases).

coupling_on
whether to implement Metropolis-coupling over temperature rungs (see also burnin_phases).

GTI_pow
the power used in the generalised thermodynamic integration method. Must be greater than 1.1.

cluster
option to pass in a cluster environment, allowing chains to be run in parallel (see package "parallel").

pb_markdown
whether to run progress bars in markdown mode, meaning they are only updated when they reach 100 being printed to markdown files.

silent
whether to suppress all console output.